requirements.txt
python-dotenv
faiss-cpu
joblib
nltk
numpy
openai
pandarallel
pandas
sentence-transformers
spacy
tqdm
